Increased Back-End Functionality
Using JWT server-stateless authentication in order to bring about authentication and authorisation. Also added middleware in order to enforce the roles of each of the roles. Edited a small error in the authentication of the Employee and Student Files as there was no code logic implemented for non-existence of user in DB. Lastly added code logic for the saving of templates by the admin users of the platform